requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
openai==1.3.5
mistralai==0.0.11
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
